Configure rabbit and EPMD to bind to an address
This change integrates against a rabbitmq cookbook feature which
is not yet merged upstream.  However, given the default code path
does not change, felt okay to get this out there.  Especially b/c
we are dependant on it.  This feature allows one to run rabbit
and clustering across a specific address.  For more details, see:
  http://tickets.opscode.com/browse/COOK-3320

Change-Id: I2a10c551cb2176dadaee8842eeadfdfc8cb93e44