Install rabbitmq-server from rabbit site for CentOS/RHEL
Fixes: bug 1282041
Implements: blueprint centos-rhel-for-ops-messaging

On CentOS/RHEL, the OS base repos don't have a package for
rabbitmq-server.  The rabbitmq cookbook by default expects to be able to
install the latest rabbitmq-server rpm from the rabbitmq site
(use_distro_version = false).  However, in our cookbooks we are overriding
'use_distro_version' to true, which works for ubuntu as it has a
rabbitmq-server package in it's base repo

This commit basically leaves the use_distro_version flag as false unless
ubuntu/suse is the platform_family, so that CentOS/RHEL can grab the rpm
from the rabbitmq site

Change-Id: I4e31cbb4957f0e696a1528fa3f7c6c80695989ac