Soften the rabbitmq version pin
With version 5.7.4 the rabbitmq cookbook should be working for us again.
However, we also learned that with the upcoming 6.x version of the
cookbook only versions >= 3.7.x of rabbitmq will be supported, which
isn't available as distro version at least for Ubuntu LTS releases yet.
So we change this pin into a soft match, allowing for updates within the
5.x stable branch, but blocking 6.x.

Change-Id: Id647f10fd5d4b69b441c4f8de8c325ec0032d9ea